refactor(e2e): move delivery lib into tests/full/prompt-delivery
- delivery helpers are suite-specific, dont belong in shared lib
- specs + lib now live together under tests/full/prompt-delivery/
